Imported HoloToolkit + Seance Stuff
Imported the packages for the holotoolkit and some of the stuff that the
other person worked on today regarding the seance stuff.  Unfortunately,
the overall seance file got corrupted, but I didn't have that much in
that one that I can't just make again.  So that's good. Just gotta get
to work now.